requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.3
soupsieve==2.7
urllib3==2.4.0
Werkzeug==3.1.3
phunspell==0.1.6
google-cloud-vision==3.10.1
google-cloud-translate==3.20.2
python-dotenv==1.1.0